requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
azure-ai-documentintelligence==1.0.1
azure-ai-ml==1.26.0
azure-common==1.1.28
azure-core==1.32.0
azure-core-tracing-opentelemetry==1.0.0b11
azure-mgmt-core==1.5.0
azure-monitor-opentelemetry==1.6.5
azure-monitor-opentelemetry-exporter==1.0.0b35
azure-storage-blob==12.25.0
azure-storage-file-datalake==12.19.0
azure-storage-file-share==12.21.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
cryptography==40.0.2
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
docx2pdf==0.1.8
exceptiongroup==1.2.2
fastapi==0.115.11
fixedint==0.1.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.45
langchain-text-splitters==0.3.6
langsmith==0.3.15
marshmallow==3.26.1
msrest==0.7.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.0.2
oauthlib==3.2.2
openai==1.66.3
opentelemetry-api==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-dbapi==0.52b0
opentelemetry-instrumentation-django==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-instrumentation-flask==0.52b0
opentelemetry-instrumentation-psycopg2==0.52b0
opentelemetry-instrumentation-requests==0.52b0
opentelemetry-instrumentation-urllib==0.52b0
opentelemetry-instrumentation-urllib3==0.52b0
opentelemetry-instrumentation-wsgi==0.52b0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0
orjson==3.10.15
packaging==24.2
propcache==0.3.0
psutil==6.1.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydash==8.0.5
PyJWT==2.10.1
PyMuPDF==1.26.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20

PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.23.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.1
strictyaml==1.7.3
tenacity==9.0.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0

pandas==2.3.1
schedule==1.1.0
pyodbc==5.2.0
APScheduler==3.11.0
openpyxl==3.1.5
